# server/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MediaTracker",
    "version": "0.2.1",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "security": [
    {
      "ApiKey": []
    },
    {
      "ApiHeader": []
    }
  ],
  "paths": {
...
    "/api/progress/by-external-id": {
      "put": {
        "tags": [
          "Progress"
        ],
        "operationId": "addByExternalId",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mediaType": {
                    "$ref": "#/components/schemas/MediaType"
                  },
                  "id": {
                    "type": "object",
                    "properties": {
                      "imdbId": {
                        "type": "string",
                        "nullable": true
                      },
                      "tmdbId": {
                        "type": "number",
                        "nullable": true
                      },
                      "audibleId": {
                        "type": "string",
                        "nullable": true
                      },
                      "igdbId": {
                        "type": "number",
                        "nullable": true
                      }
                    }
                  },
                  "seasonNumber": {
                    "type": "number",
                    "nullable": true
                  },
                  "episodeNumber": {
                    "type": "number",
                    "nullable": true
                  },
                  "action": {
                    "enum": [
                      "paused",
                      "playing",
                      null
                    ],
                    "type": "string",
                    "nullable": true
                  },
                  "progress": {
                    "type": "number",
                    "nullable": true
                  },
                  "duration": {
                    "type": "number",
                    "nullable": true
                  },
                  "device": {
                    "type": "string",
                    "nullable": true
                  }
                },
                "required": [
                  "id",
                  "mediaType"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/text": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/progress/{progressId}": {
      "delete": {
        "tags": [
          "Progress"
        ],
        "operationId": "deleteById",
        "parameters": [
          {
            "name": "progressId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/text": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
    "/api/seen": {
      "put": {
        "tags": [
          "Seen"
        ],
        "operationId": "add",
        "parameters": [
          {
            "name": "mediaItemId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "seasonId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          },
          {
            "name": "episodeId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          },
          {
            "name": "lastSeenEpisodeId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          },
          {
            "name": "lastSeenAt",
            "in": "query",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/LastSeenAt"
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          },
          {
            "name": "duration",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/text": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/seen/by-external-id": {
      "put": {
        "tags": [
          "Seen"
        ],
        "operationId": "addByExternalId",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mediaType": {
                    "$ref": "#/components/schemas/MediaType"
                  },
                  "id": {
                    "type": "object",
                    "properties": {
                      "imdbId": {
                        "type": "string",
                        "nullable": true
                      },
                      "tmdbId": {
                        "type": "number",
                        "nullable": true
                      }
                    }
                  },
                  "seasonNumber": {
                    "type": "number",
                    "nullable": true
                  },
                  "episodeNumber": {
                    "type": "number",
                    "nullable": true
                  },
                  "duration": {
                    "type": "number",
                    "nullable": true
                  }
                },
                "required": [
                  "id",
                  "mediaType"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/text": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/seen/{seenId}": {
      "delete": {
        "tags": [
          "Seen"
        ],
        "operationId": "deleteById",
        "parameters": [
          {
            "name": "seenId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/text": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/seen/": {
      "delete": {
        "tags": [
          "Seen"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "mediaItemId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "seasonId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          },
          {
            "name": "episodeId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/text": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/seen/history": {
      "get": {
        "description": "Get cross-type seen history, one entry per viewing",
        "tags": [
          "Seen"
        ],
        "operationId": "history",
        "parameters": [
          {
            "name": "mediaType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "years",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "genres",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "onlyWithUserRating",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          },
          {
            "name": "onlyWithoutUserRating",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          },
          {
            "name": "orderBy",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "date",
                "title",
                null
              ],
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc",
                null
              ],
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SeenHistoryEntry"
                      }
                    },
                    "page": {
                      "type": "number"
                    },
                    "totalPages": {
                      "type": "number"
                    },
                    "from": {
                      "type": "number"
                    },
                    "to": {
                      "type": "number"
                    },
                    "total": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "data",
                    "from",
                    "page",
                    "to",
                    "total",
                    "totalPages"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "query",
        "name": "token"
      },
      "ApiHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token"
      }
    },
    "schemas": {
...
      "MediaType": {
        "enum": [
          "audiobook",
          "book",
          "movie",
          "tv",
          "video_game"
        ],
        "type": "string"
      },
...
      "LastSeenAt": {
        "enum": [
          "custom_date",
          "now",
          "release_date",
          "unknown"
        ],
        "type": "string"
      },
      "SeenHistoryEntry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "date": {
            "type": "number",
            "nullable": true
          },
          "mediaItem": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number"
              },
              "title": {
                "type": "string"
              },
              "mediaType": {
                "type": "string"
              },
              "releaseDate": {
                "type": "string",
                "nullable": true
              },
              "genres": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "nullable": true
              },
              "poster": {
                "type": "string",
                "nullable": true
              },
              "posterSmall": {
                "type": "string",
                "nullable": true
              },
              "userRating": {
                "type": "number",
                "nullable": true
              }
            },
            "required": [
              "id",
              "mediaType",
              "title"
            ]
          },
          "episode": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number"
              },
              "seasonNumber": {
                "type": "number"
              },
              "episodeNumber": {
                "type": "number"
              },
              "title": {
                "type": "string"
              }
            },
            "required": [
              "episodeNumber",
              "id",
              "seasonNumber",
              "title"
            ],
            "nullable": true
          }
        },
        "required": [
          "id",
          "mediaItem"
        ]
      },
...
    }
  }
}